requirement.txt
opencv-python
Pillow
numpy
streamlit
